notifications

+++ api/app/models.py
@@ -728,7 +728,7 @@ class PushToken(Base):
     __tablename__ = "push_tokens"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
+    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False, index=True)
     token = Column(String, nullable=False, index=True)
     platform = Column(String, nullable=False, default="web")  # "web", "ios", "android"
     is_active = Column(Boolean, nullable=False, default=True)
